Fix issues by PR review comments
