+++ client/web/src/insights/components/InsightsViewGrid/InsightsViewGrid.tsx
@@ -1,7 +1,6 @@
 import classNames from 'classnames'
 import React, { useCallback, useMemo } from 'react'
 import { Layout as ReactGridLayout, Layouts as ReactGridLayouts, Responsive, WidthProvider } from 'react-grid-layout'
-import { noop } from 'rxjs'
 
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { isFirefox } from '@sourcegraph/shared/src/util/browserDetection'
