--- client/web/src/insights/components/InsightsViewGrid/components/insight-card/components/description /InsightCardDescription.module.scss
@@ -1,7 +1,0 @@
-.insight-description {
-    text-align: right;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    user-select: all;
-}
