+++ client/web/src/insights/components/InsightsViewGrid/components/insight-card/components/insight-card-menu/InsightCardMenu.module.scss
@@ -8,10 +8,16 @@
         list-style: none;
         background-color: var(--color-bg-1);
         border: 1px solid var(--border-color);
+
+        border-radius: 4px;
+        // TODO [VK] Copy color from dropdown menu replace by branded color
+        // class .dropdown-menu breaks Menu positioning in portal
+        box-shadow: 0 0 12px 0 rgba(0, 27, 76, 0.4);
     }
 
     &__item-button {
         width: 100%;
         text-align: start;
+        border: none;
     }
 }
