--- client/web/src/insights/components/InsightsViewGrid/components/insight-card/components/menu/InsightCardMenu.module.scss
@@ -1,23 +1,0 @@
-.menu {
-    &__panel {
-        display: flex;
-        flex-direction: column;
-        min-width: 8rem;
-        padding: 0;
-        margin: 0;
-        list-style: none;
-        background-color: var(--color-bg-1);
-        border: 1px solid var(--border-color);
-
-        border-radius: 4px;
-        // TODO [VK] Copy color from dropdown menu replace by branded color
-        // class .dropdown-menu breaks Menu positioning in portal
-        box-shadow: 0 0 12px 0 rgba(0, 27, 76, 0.4);
-    }
-
-    &__item-button {
-        width: 100%;
-        text-align: start;
-        border: none;
-    }
-}
